% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{Calculate the centralities of characters in a play.}
\usage{
centrality(play, measure)
}
\arguments{
\item{play}{The ID of a play.}

\item{measure}{"betweenness," "degree," or "eigenvector."}
}
\value{
A data table of the sorted centrality value for all characters in the \code{play}.
}
\description{
\code{centrality} takes a play (\code{play}) and a centrality measure
(\code{measure}) and returns the list of characters in the play sorted by the centrality
measure.
}
\examples{
centrality("Rom", "degree")
}
