% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_calculations.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate the cosine similarity between two vectors.}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
Cosine similarity between \code{x} and \code{y}.
}
\description{
Calculate the cosine similarity between two vectors.
}
\examples{
cosine_similarity(c(1,2,3,4), c(5,6,7,8))
}
