% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{covariance}
\alias{covariance}
\title{Calculate the covariance of a vector.}
\usage{
covariance(x, y, mode = "biased")
}
\arguments{
\item{x}{A vector.}

\item{mode}{Biased (default) or unbiased.}
}
\value{
Coariance of \code{x}.
}
\description{
Calculate the covariance of a vector.
}
\examples{
covariance(c(1,2,3,4))
covariance(c(1,2,3,4,), "unbiased")
}
