% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graph.R
\name{create_graph}
\alias{create_graph}
\title{Create a network graph of characters in a play.}
\usage{
create_graph(play)
}
\arguments{
\item{play}{The ID of a play.}
}
\value{
The network graph of all character relationships in the play.
}
\description{
\code{show_graph} takes a play (\code{play}) and creates the network graph of all character relationships in the play.
}
\examples{
create_graph("Rom")
}
