% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dot_product}
\alias{dot_product}
\title{Calculate the dot product of two vectors.}
\usage{
dot_product(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
The dot product of \code{x} and \code{y}.
}
\description{
Calculate the dot product of two vectors.
}
\examples{
dot_product(c(1,2,3,4), c(5,6,7,8))
}
