% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{entropy}
\alias{entropy}
\title{Calculate the entropy of a vector.}
\usage{
entropy(x, normalize = F)
}
\arguments{
\item{x}{A vector.}

\item{normalize}{True or false (default).}
}
\value{
The entropy of \code{x}.
}
\description{
Calculate the entropy of a vector.
}
\examples{
entropy(c(1,2,3,4))
entropy(c(1,2,3,4), T)
}
