% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_relations.R
\name{least_divergent}
\alias{least_divergent}
\title{Calculate the least divergent elements from a word.}
\usage{
least_divergent(m, word, num_results = 10)
}
\arguments{
\item{m}{A matrix.}

\item{word}{An elements of the matrix.}

\item{num_results}{Number of results to display.}
}
\value{
A data table of elements in \code{m} least divergent from \code{word}.
}
\description{
Calculate the least divergent elements from a word.
}
\examples{
least_divergent(W, "loue")
least_divergent(W, "loue", num_results = 20)
}
