% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_relations.R
\name{most_correlated}
\alias{most_correlated}
\title{Calculate the words most correlated with an input word.}
\usage{
most_correlated(m, word, num_results = 10)
}
\arguments{
\item{m}{A matrix.}

\item{word}{An entry in \code{m}.}

\item{num_results}{Number of results to display.}
}
\value{
A data table of words in \code{m} most correlated with \code{x}.
}
\description{
The correlation coefficient between vectors \code{x} and \code{y} is calculated
using \deqn{cor(x,y) = cov(x,y) / sqrt( var(x) * var(y) )} In this function, this formula is
applied between \code{x} all other word vectors in matrix \code{m}, and the resulting
correlation coefficients are sorted and returned in the form of a data table.
}
\examples{
most_correlated(W, "loue")
}
\references{
“Displaying and Summarizing Data.” \emph{Introduction to Bayesian Statistics}, by William M.
Bolstad, 2nd ed., Wiley, 2007, pp. 29–54.
}
