% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_relations.R
\name{most_similar}
\alias{most_similar}
\title{Calculate the words most similar to an input word.}
\usage{
most_similar(m, word, num_results = 10)
}
\arguments{
\item{m}{A matrix.}

\item{word}{An entry in \code{m}.}

\item{num_results}{Number of results to display.}
}
\value{
A data table of words in \code{m} most similar to \code{x}.
}
\description{
Calculate the words most similar to an input word.
}
\examples{
most_similar(W, "loue")
}
