% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{ppmi}
\alias{ppmi}
\title{Calculate positive pointwise mutual information of an array.}
\usage{
ppmi(A)
}
\arguments{
\item{A}{A matrix.}
}
\value{
PPMI of all elements in \code{A}.
}
\description{
Calculate positive pointwise mutual information of an array.
}
\examples{
ppmi(A)
}
