% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{relative_entropy}
\alias{relative_entropy}
\title{Calculate the relative entropy of two vectors.}
\usage{
relative_entropy(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector}
}
\value{
The relative entropy of \code{x} and \code{y}.
}
\description{
Calculate the relative entropy of two vectors.
}
\examples{
relative_entropy(c(1,2,3,4), c(5,6,7,8))
}
