% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_graph.R
\name{show_graph}
\alias{show_graph}
\title{Display the network graph of characters in a play.}
\usage{
show_graph(play)
}
\arguments{
\item{play}{The ID of a play.}
}
\description{
\code{show_graph} takes a play (\code{play}) and displays the network graph of all character relationships in the play.
}
\examples{
show_graph("Rom")
}
