% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{st_deviation}
\alias{st_deviation}
\title{Calculate the standard deviation of all elements in a vector.}
\usage{
st_deviation(x, mode = "biased")
}
\arguments{
\item{x}{A vector.}

\item{mode}{Biased (default) or unbiased.}
}
\value{
A vector of the standard deviation of each element in \code{x}.
}
\description{
Calculate the standard deviation of all elements in a vector.
}
\examples{
st_deviation(c(1,2,3,4))
st_deviation(c(1,2,3,4,), "unbiased")
}
