% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_relations.R
\name{top_relations}
\alias{top_relations}
\title{Returns the words most associated with a keyword.}
\usage{
top_relations(m, word, num_results = 10, method = "most_correlated")
}
\arguments{
\item{m}{A matrix.}

\item{word}{A word.}

\item{num_results}{An integer.}

\item{method}{"most_correlated," "most_similar," or "least_divergent."}
}
\value{
Top words associated with \code{word} in matrix \code{m}.
}
\description{
Returns the words most associated with a keyword.
}
\examples{
top_relations(W, "loue", 20, "most_correlated")
}
