% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{variance}
\alias{variance}
\title{Calculate the variance of a vector.}
\usage{
variance(x, mode = "biased")
}
\arguments{
\item{x}{A vector.}

\item{mode}{Biased (default) or unbiased.}
}
\value{
Variance of \code{x}.
}
\description{
Calculate the variance of a vector.
}
\examples{
variance(c(1,2,3,4))
variance(c(1,2,3,4,), "unbiased")
}
