% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_calculations.R
\name{correlation}
\alias{correlation}
\title{Calculate the correlation between two vectors.}
\usage{
correlation(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
Correlation between \code{x} and \code{y}.
}
\description{
Calculate the correlation between two vectors.
}
\examples{
correlation(c(1,2,3,4), c(1,2,3,5))
}
