% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{#' Calculate the least divergent elements from a word.
#'
#' @param m A matrix.
#' @param word An elements of the matrix.
#' @param num_results Number of results to display.
#' @return A data table of elements in \code{m} least divergent from \code{word}.
#' @examples
#' least_divergent(W, "loue")
#' least_divergent(W, "loue", num_results = 20)
least_divergent = function(m, word, num_results = 10) {
  results = apply(m, 1, relative_entropy, m[word,])
  return(sort(results, decreasing = F)[1:num_results])
}
Calculate the cosine similarity between two vectors.}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
Cosine similarity between \code{x} and \code{y}.
}
\description{
#' Calculate the least divergent elements from a word.
#'
#' @param m A matrix.
#' @param word An elements of the matrix.
#' @param num_results Number of results to display.
#' @return A data table of elements in \code{m} least divergent from \code{word}.
#' @examples
#' least_divergent(W, "loue")
#' least_divergent(W, "loue", num_results = 20)
least_divergent = function(m, word, num_results = 10) {
  results = apply(m, 1, relative_entropy, m[word,])
  return(sort(results, decreasing = F)[1:num_results])
}
Calculate the cosine similarity between two vectors.
}
\examples{
cosine_similarity(c(1,2,3,4), c(5,6,7,8))
}
