% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_calculations.R
\name{dot_product}
\alias{dot_product}
\title{Calculate the dot product of two vectors.}
\usage{
dot_product(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
The dot product of \code{x} and \code{y}.
}
\description{
Dot product, or scalar product, is calculated by taking the sum of each element
of a vector multiplied with the corresponding element in another vector:
\deqn{DP(x,y) = Σ( x(i) * y(i) )}
This function sums up the scalar multiplication of each element in the vectors, and returns
the results.
}
\examples{
dot_product(c(1,2,3,4), c(5,6,7,8))
dot_product(W["loue",], W["hate",])
}
