% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_relations.R
\name{most_correlated}
\alias{most_correlated}
\title{Calculate the words most correlated with an input word.}
\usage{
most_correlated(m, word, num_results = 10)
}
\arguments{
\item{m}{A matrix.}

\item{word}{An entry in \code{m}.}

\item{num_results}{Number of results to display.}
}
\value{
A data table of words in \code{m} most correlated with \code{x}.
}
\description{
Calculate the words most correlated with an input word.
}
\examples{
most_correlated(W, "loue")
}
