% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_relations.R
\name{most_similar}
\alias{most_similar}
\title{Calculate the words most similar to an input word.}
\usage{
most_similar(m, word, num_results = 10)
}
\arguments{
\item{m}{A matrix.}

\item{word}{An entry in \code{m}.}

\item{num_results}{Number of results to display.}
}
\value{
A data table of words in \code{m} most similar to \code{x}.
}
\description{
The similarity between two vectors is determined as follows:
\deqn{cos(x,y) = (a • b) / (||a|| ||b||)}
where (\code{a • b}) is equal to the dot product of \code{a} and \code{b}, and
(\code{||a|| ||b||}) is equal to the magnitude of \code{a} and \code{b}. This formula returns
the angle between the two vectors.
}
\examples{
most_similar(W, "loue")
}
\references{
“Word-Vectors and Search Engines.” \emph{Geometry and Meaning}, by Dominic Widdows, CSLI
Publications, 2004, pp. 132–166.
}
