% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_relations.R
\name{top_relations}
\alias{top_relations}
\title{Returns the words most associated with a keyword.}
\usage{
top_relations(m, word, method = "most correlated", num_results = 10)
}
\arguments{
\item{m}{A matrix.}

\item{word}{A word.}

\item{method}{"most correlated," "most similar," or "least divergent."}

\item{num_results}{An integer.}
}
\value{
Top words associated with \code{word} in matrix \code{m}.
}
\description{
Returns the words most associated with a keyword.
}
\examples{
top_relations(W, "loue", "most correlated", 20)
}
