% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_stats.R
\name{zscore}
\alias{zscore}
\title{Calculate the z-score of all elements in a vector.}
\usage{
zscore(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The z-score of all elements in \code{x}.
}
\description{
Calculate the z-score of all elements in a vector.
}
\examples{
zscore(c(1,2,3,4))
}
